Remove Windows x64 warnings by casting explicitly.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/7222


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@26870 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/ucnv2022.c
@@ -2091,7 +2091,7 @@ escape:
                     if(myData->version==0 && myData->key==0 && U_SUCCESS(*err) && myData->isEmptySegment) {
                         *err = U_ILLEGAL_ESCAPE_SEQUENCE;
                         args->converter->toUCallbackReason = UCNV_IRREGULAR;
-                        args->converter->toULength = toULengthBefore + (mySource - mySourceBefore);
+                        args->converter->toULength = (int8_t)(toULengthBefore + (mySource - mySourceBefore));
                     }
                 }
 
@@ -3272,7 +3272,7 @@ escape:
                     if(myData->key==0 && U_SUCCESS(*err) && myData->isEmptySegment) {
                         *err = U_ILLEGAL_ESCAPE_SEQUENCE;
                         args->converter->toUCallbackReason = UCNV_IRREGULAR;
-                        args->converter->toULength = toULengthBefore + (mySource - mySourceBefore);
+                        args->converter->toULength = (int8_t)(toULengthBefore + (mySource - mySourceBefore));
                     }
                 }
